extract EditEncryptionButton to separate component file
