+++ plugins/ros/src/components/riScDialog/RiScDialog.tsx
@@ -17,6 +17,7 @@ import ConfigEncryptionDialog from './ConfigEncryptionDialog';
 import ConfigRiscInfo from './ConfigRiscInfo';
 import { Delete } from '@material-ui/icons';
 
+
 export enum RiScDialogStates {
   Closed = 0,
   Create = 1,
