+++ plugins/ros/src/components/riScPlugin/RiScPlugin.tsx
@@ -12,12 +12,13 @@ import { getAlertSeverity } from '../../utils/utilityfunctions';
 import { useTranslationRef } from '@backstage/core-plugin-api/alpha';
 import { pluginRiScTranslationRef } from '../../utils/translations';
 import { RiScDialog, RiScDialogStates } from '../riScDialog/RiScDialog';
+import { EditEncryptionButton } from '../riScDialog/EditEncryptionButton';
 import { RiScInfo } from '../riScInfo/RiScInfo';
 import { Spinner } from '../common/Spinner';
 import { useRiScs } from '../../contexts/RiScContext';
 import { ScenarioWizardSteps } from '../../contexts/ScenarioContext';
 import { ScenarioTableWrapper } from '../scenarioTable/ScenarioTable';
-import { FeedbackDialog, EditEncryptionButton } from './FeedbackDialog.tsx';
+import { FeedbackDialog } from './FeedbackDialog.tsx';
 
 export function RiScPlugin() {
   const { t } = useTranslationRef(pluginRiScTranslationRef);
